+++ src/main/java/com/oop/dao/UserDao.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.oop.entities.AppUser;
 
 @Repository
-public interface UserDao extends CrudRepository<AppUser, Integer> {
+public interface UserDao extends CrudRepository<AppUser, Long> {
 	
 	AppUser findByUsername(String username);
